requirements.txt
dj-database-url==0.4.0
Django==1.8
django-adminactions==1.1
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.3.1
django-rest-auth==0.7.0
django-reversion==1.10.2
django-treebeard==4.0
djangorestframework==3.3.3
future==0.15.2
gunicorn==19.4.5
html2text==2014.12.29
httplib2==0.9.2
Markdown==2.6.5
oauth2client==2.0.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.1.3
pytz==2016.4
requests==2.9.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
unicodecsv==0.14.1
uritemplate==0.6
whitenoise==2.0.6
xlrd==0.9.4
xlwt-future==0.8.0
